refactor: rename future Circular and Text value indicator widgets
